refactor: sprint 39 — libinv/models.py → libinv/models/ package + Image domain
Phase 1 of the models.py split (audit item 1.1). The single-file 1800-LOC
god module becomes a package with a re-export __init__.py so every
existing import path keeps working unchanged.

Layout:
- libinv/models/__init__.py — re-exports every previously-importable
  symbol (Base from libinv.base; TimestampMixin from _base; Image
  domain from image.py; everything else still from _legacy.py).
- libinv/models/_base.py — TimestampMixin (small shared mixin).
- libinv/models/_legacy.py — the rest of the old models.py, kept as a
  single landing zone for symbols not yet split into domain files.
  Phase 2 (Sprint 40, rows 26-28) extracts Package/Vulnerability/EPSS;
  Phase 3 (Sprint 41, rows 29-33) extracts Wasp/Actionable/Repository/
  Account/Secbug/Sast. The _legacy file shrinks per phase.
- libinv/models/image.py — Image, ImagePackageAssociation, Layer,
  LatestImage (first extracted domain).

Test-mock patch targets updated where a binding-scope patch was
unavoidable: classes still living in _legacy.py must be patched at
libinv.models._legacy.X, not libinv.models.X.

- pyproject.toml: mypy override module list expanded from
  "libinv.models" to ["libinv.models", "libinv.models.*"].

Verification: 90 passed, 3 skipped, 1 xfailed (matches baseline). 216
unit tests pass. mypy libinv/models/ clean. ruff baseline preserved
(6 pre-existing issues, same as before split).
